requirements.txt
# Bittle四足歩行ロボット深層強化学習プロジェクト
# 必要なPythonパッケージリスト

# 深層強化学習フレームワーク
stable-baselines3==2.2.1
gymnasium==0.29.1

# 物理シミュレーション
pybullet==3.2.6

# 機械学習・数値計算
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# データ処理・可視化
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
plotly>=5.0.0

# 設定・ユーティリティ
pyyaml>=6.0
tqdm>=4.62.0
psutil>=5.8.0

# ログ・モニタリング
tensorboard>=2.10.0

# Jupyter Notebook
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# 画像・動画処理
opencv-python>=4.5.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.0

# 3Dモデル処理
urdfpy>=0.0.22
trimesh>=3.9.0

# 高速化
numba>=0.56.0

# 開発・テスト
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# その他ユーティリティ
pathlib2>=2.3.6
colorama>=0.4.4
rich>=12.0.0
